+++ lib/src/models.dart
@@ -7,6 +7,7 @@ export 'models/api/song.dart';
 export 'models/api/video.dart';
 export 'models/api/video_full.dart';
 export 'models/channel_filter.dart';
+export 'models/channel_video_filter.dart';
 export 'models/paginated_videos.dart';
 export 'models/search_filter.dart';
 export 'models/video_filter.dart';