+++ lib/src/utils.dart
@@ -26,9 +26,6 @@ String? concatVideoSortList(List<VideoSort> list) =>
 String? concatVideoStatusList(List<VideoStatus> list) =>
     list.isEmpty ? null : concatEnumWithCodeList(list);
 
-String concatSearchTargetList(List<SearchTarget> list) =>
-    concatEnumWithCodeList(list);
-
 String concatOrganizationList(List<Organization> list) =>
     list.map((e) => e.name).join(',');