+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 <!-- markdownlint-disable MD041 -->
 
-## 0.4.0
+## 1.0.0
 
 ### Added
 
@@ -17,7 +17,9 @@
 - **Breaking**: Renamed `HolodexClient` constructor parameter `httpClient` to `client`.
 - **Breaking**: Renamed `HolodexClient.getVideosRelatedToChannel` to `getChannelRelatedVideos`.
 - **Breaking**: Renamed `HolodexClient.getVideoMetadata` parameter `timestampComments` to `includeTimestampComments`.
-- **Breaking**: Renamed `HolodexClient.getVideoMetadata` parameter `recommendationLanguages` to `languages`.
+- **Breaking**: Renamed `HolodexClient.getVideoMetadata` parameter `recommendationLanguages` to `filterRecommendationLanguages`.
+- **Breaking**: Renamed `HolodexClient.getVideoFromId` to `getVideoById`.
+- **Breaking**: Renamed `HolodexClient.getChannelFromId` to `getChannelById`.
 - **Breaking**: Renamed `client.dart` to `holodex_client.dart`. Technically a breaking change, but only if
 implementation was imported.
 - **Breaking**: Renamed `Organization.VirtualWitchPhenomenon` to `Organization.KAMITSUBAKI`.